Written infomration for pooling information by Region.
For each region, a list of confirmed sell and confirmed buy.

Now writing information for an item.

For each item, a list of confirmed buy and sell across regions.

This will be used to determine the most profitable item, as well as the
most profitable region. That's the last step to be written.

Testing the item step... Then reorganize code.